Mol-102: QA fixes
