+++ mottak-arkiv-service/alembic/env.py
@@ -1,14 +1,8 @@
-import os
-import sys
-import logging
 from logging.config import fileConfig
-
 from sqlalchemy import engine_from_config
 from sqlalchemy import pool
-
 from alembic import context
-sys.path.insert(0, os.path.realpath(os.path.join(os.path.dirname(__file__), '..')))
-from app.db import get_url
+from app.db.database import get_url
 
 try:
     from dotenv import load_dotenv
