+++ mottak-arkiv-service/app/db/__init__.py
@@ -1,27 +1,3 @@
-import os
-import sys
-import logging
-
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-
-
-def get_url():
-    try:
-        return "%s://%s:%s@%s/%s" % (
-            os.environ["DB_DRIVER"],
-            os.environ["DB_USER"],
-            os.environ["DB_PASSWORD"],
-            os.environ["DB_HOST"],
-            os.environ["DB_NAME"],
-        )
-    except KeyError as exception:
-        logging.error(f"Environment variable not set {exception}")
-        sys.exit(1)
-
-
-def get_session():
-    engine = create_engine(get_url(), echo=True)
-    session_class = sessionmaker()
-    session_class.configure(bind=engine)
-    return session_class()
+from .schemas.mottak import *
+from .database import *
+from .repository import *
