+++ mottak-arkiv-service/app/dto/__init__.py
@@ -1,0 +1,1 @@
+from .Arkivuttrekk import *
