+++ mottak-arkiv-service/app/main.py
@@ -3,8 +3,8 @@
 from typing import List
 from fastapi import FastAPI, Depends, status
 from sqlalchemy.orm import Session
-from app.db import get_session, repository
-from app.dto.Arkivuttrekk import ArkivuttrekkOut
+from app.db import get_session, get_all_arkivuttrekk
+from app.dto import Arkivuttrekk
 
 from dotenv import load_dotenv
 load_dotenv()
@@ -33,10 +33,10 @@ async def health_check():
 
 @app.get("/arkiver",
          status_code=status.HTTP_200_OK,
-         response_model=List[ArkivuttrekkOut],
+         response_model=List[Arkivuttrekk],
          tags=["arkivuttrekk"],
          summary="Hent alle arkivuttrekk")
 def get_archives(db: Session = Depends(get_db), skip: int = 0, limit: int = 10):
-    return repository.get_all_arkivuttrekk(db, skip, limit)
+    return get_all_arkivuttrekk(db, skip, limit)
